Complete refactor plus stoppableThread class
