+++ testApp/stoppableThread.py
@@ -1,0 +1,9 @@
+from threading import Thread
+
+class StoppableThread(Thread):
+    def __init__(self):
+        Thread.__init__(self)
+        self.running = True
+
+    def stop(self):
+        self.running = False
