+++ testApp/whackerTesterHttpHandler.py
@@ -47,5 +47,5 @@ class WhackerTesterHttpHandler(BaseHTTPServer.BaseHTTPRequestHandler):
             self.send_response(404)
             self.send_header('Content-type','text/html')
             self.end_headers()
-#            self.wfile.write(get())
+            self.wfile.write(get())
             return
